+++ ruby/lib/runapi/happyhorse/resources/image_to_video.rb
@@ -21,27 +21,27 @@ module RunApi
         #
         # @param params [Hash] image-to-video parameters
         # @return [RunApi::HappyHorse::Types::CompletedImageToVideoResponse] completed task with videos
-        def run(**params)
-          task = create(**params)
-          poll_until_complete { get(task.id) }
+        def run(options: nil, **params)
+          task = create(options: options, **params)
+          poll_until_complete { get(task.id, options: options) }
         end
 
         # Create an image-to-video task.
         #
         # @param params [Hash] image-to-video parameters
         # @return [RunApi::HappyHorse::Types::ImageToVideoResponse] task creation result with id
-        def create(**params)
+        def create(options: nil, **params)
           params = compact_params(params)
           validate_params!(params)
-          request(:post, ENDPOINT, body: params)
+          request(:post, ENDPOINT, body: params, options: options)
         end
 
         # Get image-to-video task status by task ID.
         #
         # @param id [String] task ID
         # @return [RunApi::HappyHorse::Types::ImageToVideoResponse] current task status
-        def get(id)
-          request(:get, "#{ENDPOINT}/#{id}")
+        def get(id, options: nil)
+          request(:get, "#{ENDPOINT}/#{id}", options: options)
         end
 
         private